lambdas/cmr_query/requirements.txt
python-cmr
geojson-pydantic
aws-lambda-powertools
pydantic[email]
boto3
